Partial For Loop code implemented. Main change a being to use uint (and uint16) for JVM bytecode handling to avoid incorrect overflow handling for negative numbers

+++ jvm/statements/init.go
@@ -40,8 +40,8 @@ func (i InitInvocation) FillConstantsPool(pool *constantpool.ConstantPool) {
 	}
 }
 
-func (i InitInvocation) MaxStack() int {
-	max := len(i.Args)
+func (i InitInvocation) MaxStack() uint {
+	max := uint(len(i.Args))
 
 	for _, statement := range i.Args {
 		max = iMax(max, statement.MaxStack())
